Add user auth checking to Resumes and Application

+++ app/Http/Controllers/ResumeController.php
@@ -26,6 +26,7 @@ class ResumeController extends Controller
     public function index()
     {
         $resumes = Resume::query()
+            ->where('user_id', auth()->id())
             ->with(['employmentType', 'visibility'])
             ->latest()
             ->paginate(10);
@@ -68,6 +69,8 @@ class ResumeController extends Controller
      */
     public function show(Resume $resume)
     {
+        abort_if($resume->user_id !== auth()->id(), 403);
+
         $resume->load([
             'employmentType',
             'visibility',
@@ -87,6 +90,8 @@ class ResumeController extends Controller
      */
     public function edit(Resume $resume)
     {
+        abort_if($resume->user_id !== auth()->id(), 403);
+
         $resume->load('skills');
 
         $employmentTypes = EmploymentType::all();
@@ -106,6 +111,8 @@ class ResumeController extends Controller
      */
     public function update(UpdateRequest $request, Resume $resume)
     {
+        abort_if($resume->user_id !== auth()->id(), 403);
+
         $data = $request->validated();
 
         $this->service->update($resume, $data);
@@ -120,6 +127,8 @@ class ResumeController extends Controller
      */
     public function destroy(Resume $resume)
     {
+        abort_if($resume->user_id !== auth()->id(), 403);
+
         $resume->delete();
 
         return redirect()
